[#290] HapticService Dependency 구현
