+++ Qapple/Qapple/Legacy/Data/LegacyHapticService.swift
@@ -1,0 +1,22 @@
+//
+//  HapticService.swift
+//  Qapple
+//
+//  Created by 김민준 on 1/20/25.
+//
+
+import UIKit
+
+struct LegacyHapticService {
+    private init() {}
+    
+    static func notification(type: UINotificationFeedbackGenerator.FeedbackType) {
+        let generator = UINotificationFeedbackGenerator()
+        generator.notificationOccurred(type)
+    }
+    
+    static func impact(style: UIImpactFeedbackGenerator.FeedbackStyle) {
+        let generator = UIImpactFeedbackGenerator(style: style)
+        generator.impactOccurred()
+    }
+}
