+++ Qapple/Qapple/Legacy/Presentation/AnswerView/ViewModel/AnswerViewModel.swift
@@ -64,7 +64,7 @@ extension AnswerViewModel {
                 questionID: questionId
             )
             print("답변 등록 성공!")
-            HapticService.notification(type: .success)
+            LegacyHapticService.notification(type: .success)
         } catch {
             print("답변 등록 실패,,,")
         }
