+++ Qapple/Qapple/Legacy/Presentation/CommentView/LegacyCommentReportView.swift
@@ -53,7 +53,7 @@ struct LegacyCommentReportView: View {
                         Button {
                             reportType = CommentReportType.allCases[index]
                             isReportAlertPresented.toggle()
-                            HapticService.notification(type: .warning)
+                            LegacyHapticService.notification(type: .warning)
                             print("신고타입: \(reportType)")
                         } label: {
                             Text(report)
