+++ Qapple/Qapple/Legacy/Presentation/ReportView/View/LegacyReportView.swift
@@ -55,7 +55,7 @@ struct LegacyReportView: View {
                         Button {
                             reportType = LegacyReportType.allCases[index]
                             isReportAlertPresented.toggle()
-                            HapticService.notification(type: .warning)
+                            LegacyHapticService.notification(type: .warning)
                             print("신고타입: \(reportType)")
                         } label: {
                             Text(report)
