+++ Qapple/Qapple/SourceCode/Data/Service/HapticService.swift
@@ -2,21 +2,32 @@
 //  HapticService.swift
 //  Qapple
 //
-//  Created by 김민준 on 1/20/25.
+//  Created by 김민준 on 2/7/25.
 //
 
+import ComposableArchitecture
 import UIKit
 
 struct HapticService {
-    private init() {}
+    var notification: (_ type: UINotificationFeedbackGenerator.FeedbackType) -> Void
+    var impact: (_ style: UIImpactFeedbackGenerator.FeedbackStyle) -> Void
+}
+
+// MARK: - DependencyKey
+
+extension HapticService: DependencyKey {
     
-    static func notification(type: UINotificationFeedbackGenerator.FeedbackType) {
-        let generator = UINotificationFeedbackGenerator()
-        generator.notificationOccurred(type)
-    }
-    
-    static func impact(style: UIImpactFeedbackGenerator.FeedbackStyle) {
-        let generator = UIImpactFeedbackGenerator(style: style)
-        generator.impactOccurred()
+    static let liveValue = Self(
+        notification: { UINotificationFeedbackGenerator().notificationOccurred($0) },
+        impact: { UIImpactFeedbackGenerator(style: $0).impactOccurred() }
+    )
+}
+
+// MARK: - DependencyValues
+
+extension DependencyValues {
+    var hapticService: HapticService {
+        get { self[HapticService.self] }
+        set { self[HapticService.self] = newValue }
     }
 }
