+++ Qapple/Qapple/SourceCode/Feature/4.Comment/4-1.CommentList/CommentCell.swift
@@ -144,7 +144,7 @@ struct CommentCell: View {
                 // 댓글 좋아요 버튼
                 Button {
                     if !comment.isReport {
-                        HapticService.impact(style: .light)
+                        LegacyHapticService.impact(style: .light)
                         like()
                     } else {
                         self.isReportedComment.toggle()
@@ -189,7 +189,7 @@ private struct CommentDeleteButton: View {
     var body: some View {
         Button {
             delete()
-            HapticService.notification(type: .error)
+            LegacyHapticService.notification(type: .error)
         } label: {
             ZStack {
                 Color.delete
