+++ Qapple/Qapple/SourceCode/Feature/4.Comment/4-1.CommentList/CommentView.swift
@@ -144,7 +144,7 @@ private struct AddCommentView: View {
                 .padding(.vertical, 12)
             
             Button {
-                    HapticService.notification(type: .success)
+                    LegacyHapticService.notification(type: .success)
                     store.send(.uploadCommentButtonTapped)
             } label: {
                 Image(systemName: "paperplane")
