+++ Qapple/Qapple/SourceCode/Feature/4.Comment/4-2.CommentReport/CommentReportView.swift
@@ -47,7 +47,7 @@ struct CommentReportView: View {
                     ForEach(Array(self.reportList.enumerated()), id: \.offset) { index, report in
                         Button {
                             store.send(.reportListItemTapped(index))
-                            HapticService.notification(type: .warning)
+                            LegacyHapticService.notification(type: .warning)
                         } label: {
                             Text(report)
                                 .font(.pretendard(.medium, size: 16))
