+++ Qapple/Qapple/SourceCode/Feature/5.Profile/5-1.Profile/ProfileView.swift
@@ -180,7 +180,7 @@ private struct AccountSection: View {
                 icon: .signOutIcon,
                 tapAction: {
                     store.send(.logOutButtonTapped)
-                    HapticService.notification(type: .warning)
+                    LegacyHapticService.notification(type: .warning)
                 }
             )
             
@@ -190,7 +190,7 @@ private struct AccountSection: View {
                 isDistructive: true,
                 tapAction: {
                     store.send(.resignButtonTapped)
-                    HapticService.notification(type: .warning)
+                    LegacyHapticService.notification(type: .warning)
                 }
             )
         }
